test_classes: Extract home view helpers for reuse.
